Изменен тип атрибута класса Student ID Он должен быть long - это упрощает взаимодействие с БД

+++ app/src/main/java/com/example/artemij/ppplus/EditActivity.java
@@ -63,7 +63,7 @@ public class EditActivity extends AppCompatActivity {
         if (type.equals("NEW")) {
             setUpNewEnvironment();
         } else if (type.equals("EDIT")) {
-            int studID = i.getIntExtra("StudentID", -1);
+            long studID = i.getLongExtra("StudentID", -1);
             if (studID != -1) {
                 studEdit = dbStudent.getStudent(studID);
                 setUpEditEnvironment();
